:sparkles: Download all aether xml

+++ defs/json_data.py
@@ -43,9 +43,11 @@ async def get_draws(gender_boy: bool = True) -> List["MyDraw"]:
 
 
 async def save_raw_jsons():
+    path = FILE_PATH_BOY if GENSHIN_GENDER_BOY else FILE_PATH_GIRL
+    file_path = path / "my_draws.json"
+    if file_path.exists():
+        return
     draws = await get_my_draws()
     data = [draw.dict() for draw in draws if draw.take_picture]
-    path = FILE_PATH_BOY if GENSHIN_GENDER_BOY else FILE_PATH_GIRL
-    with open(path / "my_draws.json", "w", encoding="utf-8") as f:
+    with open(file_path, "w", encoding="utf-8") as f:
         json.dump(data, f, ensure_ascii=False, indent=4)
-    print("Done")
